requirements.txt
click==8.0.4
deepface==0.0.79
dlib==19.22.1
imageio==2.27.0
matplotlib==3.7.1
numpy==1.22.4
opencv_python_headless==4.7.0.72
Pillow==9.4.0
PyQt5==5.15.9
PyQt5_sip==12.11.0
Requests==2.30.0
scikit_learn==1.2.2
scipy==1.9.1
setuptools==65.6.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.62.2
training==0.0.1
transformers==4.27.4
